# contracts/openapi/safety_kernel.yaml
...
openapi: 3.1.0
info:
  title: QantmOrchstrtr-RSI Safety Kernel
  version: 1.0.0
  description: |
    Authorization, attestation, and audit-chain service for sensitive
    operations across the QantmOrchstrtr-RSI system. The kernel is the
    only component authorized to mint signed tokens for actions like
    `sio_run_cycles`, RSI cycle persistence, and approval decisions.

    Caller roles (set by middleware via `x-api-key`):
    - `worker` — primary caller; authorized to request tokens for
      RSI/SIO/training actions
    - `api` — REST API; restricted to governance-only actions per
      `is_api_action_allowed()` allowlist
    - `operator` — human approvals/rejections only
  contact:
    name: ARYA Labs Safety Kernel team
    url: https://github.com/ARYA-Labs-PBC/QantmOrchstrtr-RSI
  license:
    name: Proprietary
servers:
  - url: http://safety-kernel:9000
    description: Internal docker network (production)
  - url: http://localhost:9001
    description: Host port mapping (local development)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Pre-shared key set in `.env` per role:
        - `QORCH_KERNEL_API_KEY_WORKER`
        - `QORCH_KERNEL_API_KEY_API`
        - `QORCH_KERNEL_API_KEY_OPERATOR`
  schemas:
...
    AuthorizeRequest:
      type: object
      required: [action, run_id, subject, params_fingerprint]
      properties:
        action:
          type: string
          minLength: 1
          description: Sensitive action to authorize (e.g., `sio_run_cycles`)
        run_id:
          type: string
          minLength: 1
          description: Run identifier bound into the token
        subject:
          type: string
          minLength: 1
          description: Worker identity / subject requesting authorization
        params_fingerprint:
          type: string
          minLength: 1
          pattern: ^[0-9a-f]{64}$
          description: SHA-256 fingerprint of parameters (stable JSON)
        params:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional params used for policy decisions (fingerprinted)
        ttl_s:
          type: integer
          minimum: 1
          maximum: 3600
          nullable: true
          description: Requested TTL seconds (kernel will clamp)
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional metadata for audit (not used in token binding)
    AuthorizeResponse:
      type: object
      required: [ok, token, token_sha256, claims]
      properties:
        ok:
          type: boolean
          enum: [true]
        token:
          type: string
          description: Ed25519-signed compact JWT token
        token_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
          description: SHA-256 of the token string (for audit)
        claims:
          type: object
          additionalProperties: true
          description: Decoded claims (action, run_id, subject, exp, etc.)
...
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Stable machine-readable error code
        reason:
          type: string
          description: Optional human-readable reason

    # ------------------------------------------------------------------
    # /policy/* (ADR-018, ARY-2028) — module-load / exec / compile
    # authorization. Slice 1 freezes ONLY PolicyModuleAuthorize{Request,
    # AllowResponse, DenyResponse}. The /policy/module/register,
    # /policy/audit-event, /policy/module/{path}/status endpoints are
    # listed in `paths:` with a DRAFT description and 501-only
    # responses; slice 2 fixes their request/response schemas.
    # ------------------------------------------------------------------
...
  responses:
    Unauthorized:
      description: Missing / invalid `x-api-key`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        Role-based forbidden. `caller_role_forbidden` (unknown role) or
        `caller_role_not_operator` (operator-only endpoint) or
        `api_action_forbidden` (API role attempting non-governance action).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /kernel/v1/authorize:
    post:
      operationId: postAuthorize
      summary: Authorize a sensitive action (signed token)
      description: |
        Returns a signed Ed25519 token authorizing the specified action
        for the specified subject. The token's `exp` claim is the kernel's
        clamped TTL; clients MUST honour it. The audit chain records
        the authorization decision before the token is returned.
      tags: [kernel]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Validation error (Pydantic / FastAPI default body)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...